Fix en passant square updating

+++ src/board/legality/mod.rs
@@ -9,9 +9,7 @@ enum GameState {
     Win(Color),
     StillInProgress,
 }
-
 pub mod getting_out_of_check;
-
 pub const MOVE_AVAILABILITY_VALIDATORS: [fn(usize, &mut Board) -> bool; 7] = [
     pawn::can_be_moved,
     rook::can_be_moved,
@@ -115,7 +113,7 @@ impl Board {
     fn is_material_sufficient_to_checkmate(&mut self) -> bool {
         #[inline]
         fn get_square_color(location: usize) -> Color {
-            ((location >> 3) & 1 == location & 1 && location & 1 == 0) as u8
+            (location >> 3) & 1 == location & 1 && location & 1 == 0
         }
 
         match self.state.pieces_count {
